change backlight property type to long

+++ common/xfpm-brightness.h
@@ -51,20 +51,20 @@ XfpmBrightness       	       *xfpm_brightness_new             (void);
 gboolean			xfpm_brightness_setup 		(XfpmBrightness *brightness);
 
 gboolean			xfpm_brightness_up		(XfpmBrightness *brightness,
-								 gint *new_level);
+								 glong *new_level);
 
 gboolean			xfpm_brightness_down		(XfpmBrightness *brightness,
-								 gint *new_level);
+								 glong *new_level);
 
 gboolean			xfpm_brightness_has_hw 		(XfpmBrightness *brightness);
 
 gint 				xfpm_brightness_get_max_level   (XfpmBrightness *brightness);
 
 gboolean			xfpm_brightness_get_level	(XfpmBrightness *brightness,
-								 gint *level);
+								 glong *level);
 
 gboolean			xfpm_brightness_set_level	(XfpmBrightness *brightness,
-								 gint level);
+								 glong level);
 
 gboolean			xfpm_brightness_dim_down	(XfpmBrightness *brightness);
